+++ program-libs/compressed-account/src/instruction_data/mod.rs
@@ -1,5 +1,6 @@
 pub mod compressed_proof;
 pub mod cpi_context;
 pub mod data;
+pub mod insert_into_queues;
 pub mod invoke_cpi;
 pub mod zero_copy;
