+++ programs/account-compression/src/processor/insert_leaves.rs
@@ -1,5 +1,5 @@
 use anchor_lang::prelude::*;
-use light_compressed_account::insert_into_queues::AppendLeavesInput;
+use light_compressed_account::instruction_data::insert_into_queues::AppendLeavesInput;
 
 use crate::{context::AcpAccount, errors::AccountCompressionErrorCode};
