+++ programs/account-compression/src/processor/insert_nullifiers.rs
@@ -1,6 +1,6 @@
 use anchor_lang::prelude::*;
 use light_batched_merkle_tree::queue::BatchedQueueAccount;
-use light_compressed_account::insert_into_queues::InsertNullifierInput;
+use light_compressed_account::instruction_data::insert_into_queues::InsertNullifierInput;
 use num_bigint::BigUint;
 
 use crate::{
