+++ programs/system/src/processor/create_inputs_cpi_data.rs
@@ -2,8 +2,10 @@ use account_compression::{context::AcpAccount, errors::AccountCompressionErrorCo
 use anchor_lang::prelude::*;
 use light_compressed_account::{
     hash_to_bn254_field_size_be,
-    insert_into_queues::{InsertIntoQueuesInstructionDataMut, InsertNullifierInput},
-    instruction_data::zero_copy::ZPackedCompressedAccountWithMerkleContext,
+    instruction_data::{
+        insert_into_queues::{InsertIntoQueuesInstructionDataMut, InsertNullifierInput},
+        zero_copy::ZPackedCompressedAccountWithMerkleContext,
+    },
 };
 use light_hasher::{Hasher, Poseidon};
